--- a/Neos.Flow/Classes/Core/Booting/Scripts.php
+++ b/Neos.Flow/Classes/Core/Booting/Scripts.php
@@ -728,9 +728,9 @@ class Scripts
     /**
      * Check if the old fallback classloader should be used.
      *
-     * The old class loader is not used only in one case
-     * * the environment variable "FLOW_ONLY_COMPOSER_LOADER" was set
-     * * not in a testing context
+     * The old class loader is used only in the cases:
+     * * the environment variable "FLOW_ONLY_COMPOSER_LOADER" is not set or false
+     * * in a testing context
      *
      * @param Bootstrap $bootstrap
      * @return bool