Currently the composer class loader is only used as a fallback to our own,
but especially if the optimized loader is used the composer one is much
faster.

On systems in which all packages/classes are registered correctly via
autoload statements in ``composer.json`` files using our own class loader
only for proxy classes can bring an substantial boost in performance for
every request.

In order to enable this feature you need to set an environment variable
``FLOW_ONLY_COMPOSER_LOADER=1``. Please test carefully if that breaks due
to your autoload configuration not being fully composer ready.

Additionally it is recommended to use the optimized composer loader by
calling ``composer dumpautoload -o``.

While not breaking in itself this change deprecates using our class loader
for anything else than proxy classes. In practice this means you should
always enable composer auto loader only by using above mentioned environment
variable. At least make sure that your projects work with this
enabled.

We will drop the variable and make this the default behavior in the next
major version of Flow (version 5.0) which means only classes that are
correctly added to composer (loaded) packages with autoload configuration
are being loaded correctly.