@kdambekalns 2017-04-03
Is that flag supposed to stay? Could we invert the default? Use our loader as a fallback to it's faster but "unbreakable"?