@kitsunet 2017-04-03
Ok so the following variants exist:

1) We split our loader between proxy and rest, so we can sandwich the composer loader:
- proxy loader
- composer loader
- flow class loader

That would work fine and I could do it, it's a bit bigger change though. We could then get rid of the flag of course. Has the drawback that our loader would still be initialized with a lot of information that is in theory superfluous (but we can't know that), so it eats up a bit of the performance gain.

2) My plan was to deprecate any non composer loaded classes and remove the flag and the respective parts of our loader on next major.