@kitsunet 2017-04-03
Just for clarification, simply reversing the order is impossible as our proxy loading MUST come first.