@albe 2017-04-03
> Has the drawback that our loader would still be initialized with a lot of information that is in theory superfluous (but we can't know that), so it eats up a bit of the performance gain.

Can we just keep the env var flag as an further optimization then and keep moving towards your goal in 2., using the flag as the "check that everything works when the flow class loader is disabled"?