@kitsunet 2017-04-04
Ok, updated commit (should also fix tests) and added hint about deprecation to commit message.