@albe 2017-04-28
Sorry, didn't come around to actually test it in our project and will also be on vacation the next 2 weeks now