@albe 2017-06-18
Looks good on a first try with my local Neos.EventSourcing project. Could not yet try our work projects, since the servers can currently not easily be updated to PHP7 and hence Flow 4.

We really need to document the fact that non-composer-installed packages will break very well. Will try to get some more tests done.