@robertlemke 2017-06-22
> but especially if the optimized loader is used the composer one is much
faster.

which speed gain did you measure, for example based on the Neos demo site?