@dfeyer 2017-11-04
Can we merge this one, it just make non sense to have two class loader. And if we can slim down our code base, and maybe improve performance, it's just a win win. 

Fine also to deprecate our loader, and remove it in the next major. Now the community is bit more used to composer usages, and it's fairly easy to adapt project than can be broken by this the change in the next major. We can even write a blog post on the subjet. The problem concern mainly packages that are include in the distribution. It's easy to transform those distribution with local composer sources.