@kitsunet 2017-11-12
WITH patch:

```
Lifting the server siege...
Transactions:		        6858 hits
Availability:		      100.00 %
Elapsed time:		       59.10 secs
Data transferred:	        4.09 MB
Response time:		        0.21 secs
Transaction rate:	      116.04 trans/sec
Throughput:		        0.07 MB/sec
Concurrency:		       24.93
Successful transactions:        6858
Failed transactions:	           0
Longest transaction:	        0.52
Shortest transaction:	        0.07
```


WITHOUT patch:
```
Lifting the server siege...
Transactions:		        6467 hits
Availability:		      100.00 %
Elapsed time:		       59.14 secs
Data transferred:	        3.85 MB
Response time:		        0.23 secs
Transaction rate:	      109.35 trans/sec
Throughput:		        0.07 MB/sec
Concurrency:		       24.94
Successful transactions:        6467
Failed transactions:	           0
Longest transaction:	        0.74
Shortest transaction:	        0.04
```

pretty arbitrary though. Can probably go the other way as well. Would be interesting to benchmark on a non SSD system for a change. I _think_ that might show more difference.