@albe 2017-11-18
I think that's already good enough of an improvement to justify this, let alone the simplification of our code base in the mid term.

I'm unsure about if this should land in LTS, because the "breakiness" of this is now only on non-composer installed packages. Should we be strict about b/c with this?