@kitsunet 2017-12-06
@albe it is fully backwards compatible right now, especially as long as you DON'T set the env. If you set it, ofc it might break, but you actively have to go in and set that env.