@kitsunet 2017-12-11
It deprecates the old way and we always marked these kinds of changes "breaking" to warn users that big change is happening in a future version.