@kitsunet 2017-12-15
Yes, well it is as easy as configuring functional test classes in autoload. But now the interesting question, do we actually want to do that. I would rather suggest the following: 

We add a testing autoloader that goes through composer.json of installed packages and uses the require-dev to build a map just like our autoloader does now. That way tests would work fine with autoload-dev configured in local packages and we can move forward with this.