@kitsunet 2017-12-15
Here we go. This is not exactly the final state yet. But better than what we have now with the option of extra performance by usign the flag.