@kitsunet 2017-12-15
Not sure what those styleci fails are. Nto exactly code I even touched.