@kitsunet 2017-12-15
There we go, merged master into it which makes styleci happy again.

Would merge this weekend if no one else blocks it.